+++ examples/clone.js
@@ -2,7 +2,6 @@ const P2PCommons = require('..') // liberate science constructor function
 const tempy = require('tempy')
 const commons = new P2PCommons({
   verbose: true,
-  persist: false,
   baseDir: tempy.directory()
 })
 
@@ -18,8 +17,9 @@ process.once('SIGINT', () => commons.destroy())
     throw new Error('missing arg: external module url')
   }
   console.time('clone')
-  const { rawJSON } = await commons.clone(externalContentUrl, version, false)
+  const { rawJSON } = await commons.clone(externalContentUrl, version)
   console.timeEnd('clone')
   console.log({ rawJSON })
+  console.log({ baseDir: commons.baseDir })
   await commons.destroy()
 })()